third_party/blink/web_tests/fast/forms/select/options-default-white-space.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<select>
  <option id=option value="hello">Hello</option>
</select>

<script>
test(() => {
  assert_equals(
    window.getComputedStyle(option)['white-space'],
    'nowrap');
}, `<option>'s user agent 'white-space' should be 'nowrap'.`);
</script>
